requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
attrs==25.4.0
backoff==2.2.1
backports.functools-lru-cache==2.0.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.7
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
durationpy==0.10
fastapi==0.109.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
gTTS==2.5.4
h11==0.16.0
h3==4.4.1
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
indic-transliteration==2.3.75
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.8.2
kagglesdk==0.1.13
keras==3.13.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.1.2
langchain-core==1.1.1
langchain-openai==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
libclang==18.1.1
librt==0.7.4
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.14.0
neo4j==5.16.0
networkx==3.6
numpy==2.2.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.9.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
posthog==5.4.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytokens==0.3.0
pytz==2024.1
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
roman==5.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.35.1
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
text-unidecode==1.3
tf_keras==2.20.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
Wikipedia-API==0.8.1
wrapt==2.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
